raspi/requirements.txt
paho-mqtt
bleak
pyyaml
